refactor(tabs): Barrel file renaming
