+++ projects/igniteui-angular/src/lib/tabs/public_api.ts
